refactor(setup): move GitHub Actions env persistence into its own module
Route the GITHUB_ACTIONS, GITHUB_ENV, and GITHUB_PATH lookups through the
EnvVarOs capability seam instead of calling std::env::var_os in library
code, per the "Reading process state" section of the pacquet code style
guide. Tests can now drive the inside-a-workflow and outside-a-workflow
branches with a fake provider rather than only the write helpers.

Put the pacquet side in cli_args/setup/github_actions_env.rs with its unit
tests in the external tests submodule, and the TypeScript side in
setup/githubActionsEnv.ts, so setup.rs and setup.ts are back to the
install-and-extend-PATH flow alone.
