+++ pnpm/crates/cli/src/cli_args/setup.rs
@@ -5,26 +5,17 @@
 //! `PNPM_HOME` plus `$PNPM_HOME/bin` are added to the user's environment
 //! (the shell rc file on POSIX, the registry on Windows).
 
+mod github_actions_env;
 mod path_extender;
 
 use clap::Args;
-use derive_more::{Display, Error};
-use miette::{Context, Diagnostic, IntoDiagnostic};
+use miette::{Context, IntoDiagnostic};
 use pacquet_config::{Host, PNPM_VERSION, default_pnpm_home_dir};
 use pacquet_reporter::{LogEvent, LogLevel, PnpmLog, Reporter};
 use path_extender::{
     AddDirToEnvPathOpts, AddingPosition, ConfigFileChangeType, ConfigReport, PathExtenderReport,
 };
-use std::{
-    ffi::OsStr,
-    fs::{self, File, OpenOptions},
-    io::{Read, Seek, SeekFrom, Write},
-    path::{Path, PathBuf},
-    process::Command,
-};
-
-#[cfg(unix)]
-use std::os::unix::fs::OpenOptionsExt;
+use std::{fs, path::Path, process::Command};
 
 #[derive(Debug, Args)]
 pub struct SetupArgs {
@@ -57,7 +48,7 @@ fn handler<Reporter: self::Reporter + 'static>(force: bool, dir: &Path) -> miett
     // the self-install subprocess's `PATH` or the alias-script writes.
     path_extender::validate_pnpm_home_dir(&pnpm_home_dir)?;
     let bin_dir = pnpm_home_dir.join("bin");
-    validate_github_actions_environment_file_values(&pnpm_home_dir, &bin_dir)?;
+    github_actions_env::validate_persisted_values::<Host>(&pnpm_home_dir, &bin_dir)?;
 
     let exec_path = std::env::current_exe()
         .into_diagnostic()
@@ -78,170 +69,9 @@ fn handler<Reporter: self::Reporter + 'static>(force: bool, dir: &Path) -> miett
             position: AddingPosition::Start,
         },
     )?;
-    persist_github_actions_environment::<Reporter>(dir, &pnpm_home_dir, &bin_dir);
+    github_actions_env::persist::<Reporter, Host>(dir, &pnpm_home_dir, &bin_dir);
     remove_legacy_homedir_shims(&pnpm_home_dir);
     Ok(render_setup_output(&report))
-}
-
-/// `GITHUB_ENV` and `GITHUB_PATH` are line-oriented, so a line break in a
-/// persisted value would append attacker-chosen records to the environment of
-/// every later step in the workflow job.
-#[derive(Debug, Display, Error, Diagnostic)]
-#[display("{name} cannot contain newline or NUL characters")]
-#[diagnostic(code(ERR_PNPM_BAD_GITHUB_ACTIONS_ENVIRONMENT_VALUE))]
-struct BadGitHubActionsEnvironmentValue {
-    name: &'static str,
-}
-
-fn validate_github_actions_environment_file_values(
-    pnpm_home_dir: &Path,
-    bin_dir: &Path,
-) -> miette::Result<()> {
-    if !should_persist_github_actions_environment_files() {
-        return Ok(());
-    }
-    validate_github_actions_environment_file_value("PNPM_HOME", pnpm_home_dir)?;
-    validate_github_actions_environment_file_value("pnpm setup bin directory", bin_dir)
-}
-
-fn validate_github_actions_environment_file_value(
-    name: &'static str,
-    value: &Path,
-) -> miette::Result<()> {
-    if value.to_string_lossy().contains(['\n', '\r', '\0']) {
-        return Err(BadGitHubActionsEnvironmentValue { name }.into());
-    }
-    Ok(())
-}
-
-fn persist_github_actions_environment<Reporter: self::Reporter>(
-    prefix_dir: &Path,
-    pnpm_home_dir: &Path,
-    bin_dir: &Path,
-) {
-    let github_env = std::env::var_os("GITHUB_ENV").map(PathBuf::from);
-    let github_path = std::env::var_os("GITHUB_PATH").map(PathBuf::from);
-    persist_github_actions_environment_to_files::<Reporter>(
-        prefix_dir,
-        is_github_actions(),
-        pnpm_home_dir,
-        bin_dir,
-        github_env.as_deref(),
-        github_path.as_deref(),
-    );
-}
-
-fn persist_github_actions_environment_to_files<Reporter: self::Reporter>(
-    prefix_dir: &Path,
-    is_github_actions: bool,
-    pnpm_home_dir: &Path,
-    bin_dir: &Path,
-    github_env: Option<&Path>,
-    github_path: Option<&Path>,
-) {
-    if !is_github_actions {
-        return;
-    }
-    write_github_actions_environment_files::<Reporter>(
-        prefix_dir,
-        pnpm_home_dir,
-        bin_dir,
-        github_env,
-        github_path,
-    );
-}
-
-fn should_persist_github_actions_environment_files() -> bool {
-    is_github_actions()
-        && (std::env::var_os("GITHUB_ENV").is_some() || std::env::var_os("GITHUB_PATH").is_some())
-}
-
-fn is_github_actions() -> bool {
-    std::env::var_os("GITHUB_ACTIONS").is_some_and(|value| value == OsStr::new("true"))
-}
-
-fn write_github_actions_environment_files<Reporter: self::Reporter>(
-    prefix_dir: &Path,
-    pnpm_home_dir: &Path,
-    bin_dir: &Path,
-    github_env: Option<&Path>,
-    github_path: Option<&Path>,
-) {
-    if let Some(github_env) = github_env {
-        append_github_actions_environment_file::<Reporter>(
-            prefix_dir,
-            "GITHUB_ENV",
-            github_env,
-            &format!("PNPM_HOME={}", pnpm_home_dir.display()),
-        );
-    }
-    if let Some(github_path) = github_path {
-        append_github_actions_environment_file::<Reporter>(
-            prefix_dir,
-            "GITHUB_PATH",
-            github_path,
-            &format!("{}", bin_dir.display()),
-        );
-    }
-}
-
-fn append_github_actions_environment_file<Reporter: self::Reporter>(
-    prefix_dir: &Path,
-    target_name: &str,
-    path: &Path,
-    line: &str,
-) {
-    if let Err(err) = append_existing_regular_file(path, line) {
-        warn::<Reporter>(
-            prefix_dir,
-            &format!(
-                "Failed to write GitHub Actions environment file {target_name} ({}): {err}",
-                path.display(),
-            ),
-        );
-    }
-}
-
-fn append_existing_regular_file(path: &Path, line: &str) -> std::io::Result<()> {
-    match fs::symlink_metadata(path) {
-        Ok(metadata) if metadata.file_type().is_file() => {}
-        Ok(_) => return Ok(()),
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => return Ok(()),
-        Err(err) => return Err(err),
-    }
-    let mut file = match open_existing_file_for_append(path) {
-        Ok(file) => file,
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => return Ok(()),
-        Err(err) => return Err(err),
-    };
-    if !file.metadata()?.file_type().is_file() {
-        return Ok(());
-    }
-    write_line_to_file(&mut file, line)?;
-    Ok(())
-}
-
-fn open_existing_file_for_append(path: &Path) -> std::io::Result<File> {
-    let mut options = OpenOptions::new();
-    options.read(true).append(true);
-    #[cfg(unix)]
-    options.custom_flags(libc::O_NOFOLLOW);
-    options.open(path)
-}
-
-fn write_line_to_file(file: &mut File, line: &str) -> std::io::Result<()> {
-    let mut output = String::new();
-    if file.metadata()?.len() > 0 {
-        let mut last_byte = [0];
-        file.seek(SeekFrom::End(-1))?;
-        file.read_exact(&mut last_byte)?;
-        if last_byte[0] != b'\n' {
-            output.push('\n');
-        }
-    }
-    output.push_str(line);
-    output.push('\n');
-    file.write_all(output.as_bytes())
 }
 
 /// Install the CLI as a global package using `pnpm add -g file:<dir>`,
@@ -397,14 +227,6 @@ fn info<Reporter: self::Reporter>(prefix: &str, message: &str) {
         level: LogLevel::Info,
         message: message.to_string(),
         prefix: prefix.to_string(),
-    }));
-}
-
-fn warn<Reporter: self::Reporter>(prefix: &Path, message: &str) {
-    Reporter::emit(&LogEvent::Pnpm(PnpmLog {
-        level: LogLevel::Warn,
-        message: message.to_string(),
-        prefix: prefix.to_string_lossy().into_owned(),
     }));
 }
