+++ pnpm/crates/cli/src/cli_args/setup/tests.rs
@@ -4,13 +4,10 @@
 
 use super::{
     ConfigFileChangeType, ConfigReport, LEGACY_HOME_DIR_SHIM_NAMES, PathExtenderReport,
-    create_alias_scripts, persist_github_actions_environment_to_files, remove_legacy_homedir_shims,
-    render_setup_output, validate_github_actions_environment_file_value,
-    write_github_actions_environment_files,
+    create_alias_scripts, remove_legacy_homedir_shims, render_setup_output,
 };
-use pacquet_reporter::{LogEvent, LogLevel, PnpmLog, Reporter, SilentReporter};
 use pretty_assertions::assert_eq;
-use std::{path::PathBuf, sync::Mutex};
+use std::path::PathBuf;
 
 fn report(change_type: ConfigFileChangeType, old: &str, new: &str) -> PathExtenderReport {
     PathExtenderReport {
@@ -99,220 +96,4 @@ fn remove_legacy_homedir_shims_tolerates_missing_files() {
     // not treat absent files as an error.
     let dir = tempfile::tempdir().expect("create temp dir");
     remove_legacy_homedir_shims(dir.path());
-}
-
-#[test]
-fn github_actions_environment_files_receive_home_and_bin() {
-    let dir = tempfile::tempdir().expect("create temp dir");
-    let pnpm_home_dir = dir.path().join("pnpm-home");
-    let bin_dir = pnpm_home_dir.join("bin");
-    let github_env = dir.path().join("github-env");
-    let github_path = dir.path().join("github-path");
-    std::fs::write(&github_env, "").expect("create github env");
-    std::fs::write(&github_path, "").expect("create github path");
-
-    write_github_actions_environment_files::<SilentReporter>(
-        dir.path(),
-        &pnpm_home_dir,
-        &bin_dir,
-        Some(&github_env),
-        Some(&github_path),
-    );
-
-    assert_eq!(
-        std::fs::read_to_string(github_env).expect("read github env"),
-        format!("PNPM_HOME={}\n", pnpm_home_dir.display()),
-    );
-    assert_eq!(
-        std::fs::read_to_string(github_path).expect("read github path"),
-        format!("{}\n", bin_dir.display()),
-    );
-}
-
-#[test]
-fn github_actions_environment_files_start_new_records_after_existing_content() {
-    let dir = tempfile::tempdir().expect("create temp dir");
-    let pnpm_home_dir = dir.path().join("pnpm-home");
-    let bin_dir = pnpm_home_dir.join("bin");
-    let github_env = dir.path().join("github-env");
-    let github_path = dir.path().join("github-path");
-    std::fs::write(&github_env, "EXISTING=value").expect("create github env");
-    std::fs::write(&github_path, "/existing/bin").expect("create github path");
-
-    write_github_actions_environment_files::<SilentReporter>(
-        dir.path(),
-        &pnpm_home_dir,
-        &bin_dir,
-        Some(&github_env),
-        Some(&github_path),
-    );
-
-    assert_eq!(
-        std::fs::read_to_string(github_env).expect("read github env"),
-        format!("EXISTING=value\nPNPM_HOME={}\n", pnpm_home_dir.display()),
-    );
-    assert_eq!(
-        std::fs::read_to_string(github_path).expect("read github path"),
-        format!("/existing/bin\n{}\n", bin_dir.display()),
-    );
-}
-
-#[test]
-fn github_actions_environment_files_are_ignored_outside_github_actions() {
-    let dir = tempfile::tempdir().expect("create temp dir");
-    let pnpm_home_dir = dir.path().join("pnpm-home");
-    let bin_dir = pnpm_home_dir.join("bin");
-    let github_env = dir.path().join("github-env");
-    let github_path = dir.path().join("github-path");
-    std::fs::write(&github_env, "").expect("create github env");
-    std::fs::write(&github_path, "").expect("create github path");
-
-    persist_github_actions_environment_to_files::<SilentReporter>(
-        dir.path(),
-        false,
-        &pnpm_home_dir,
-        &bin_dir,
-        Some(&github_env),
-        Some(&github_path),
-    );
-
-    assert_eq!(std::fs::read_to_string(github_env).expect("read github env"), "");
-    assert_eq!(std::fs::read_to_string(github_path).expect("read github path"), "");
-}
-
-#[test]
-fn github_actions_environment_files_are_written_independently() {
-    let dir = tempfile::tempdir().expect("create temp dir");
-    let pnpm_home_dir = dir.path().join("pnpm-home");
-    let bin_dir = pnpm_home_dir.join("bin");
-    let github_env = dir.path().join("github-env");
-    let github_path = dir.path().join("github-path");
-    std::fs::write(&github_env, "").expect("create github env");
-    std::fs::write(&github_path, "").expect("create github path");
-
-    persist_github_actions_environment_to_files::<SilentReporter>(
-        dir.path(),
-        true,
-        &pnpm_home_dir,
-        &bin_dir,
-        Some(&github_env),
-        None,
-    );
-    persist_github_actions_environment_to_files::<SilentReporter>(
-        dir.path(),
-        true,
-        &pnpm_home_dir,
-        &bin_dir,
-        None,
-        Some(&github_path),
-    );
-
-    assert_eq!(
-        std::fs::read_to_string(github_env).expect("read github env"),
-        format!("PNPM_HOME={}\n", pnpm_home_dir.display()),
-    );
-    assert_eq!(
-        std::fs::read_to_string(github_path).expect("read github path"),
-        format!("{}\n", bin_dir.display()),
-    );
-}
-
-#[test]
-fn github_actions_environment_files_skip_non_regular_targets() {
-    let dir = tempfile::tempdir().expect("create temp dir");
-    let pnpm_home_dir = dir.path().join("pnpm-home");
-    let bin_dir = pnpm_home_dir.join("bin");
-    let github_env = dir.path().join("github-env-dir");
-    let github_path = dir.path().join("github-path");
-    std::fs::create_dir(&github_env).expect("create github env dir");
-    std::fs::write(&github_path, "").expect("create github path");
-
-    persist_github_actions_environment_to_files::<SilentReporter>(
-        dir.path(),
-        true,
-        &pnpm_home_dir,
-        &bin_dir,
-        Some(&github_env),
-        Some(&github_path),
-    );
-
-    assert!(github_env.is_dir());
-    assert_eq!(
-        std::fs::read_to_string(github_path).expect("read github path"),
-        format!("{}\n", bin_dir.display()),
-    );
-}
-
-#[test]
-fn github_actions_environment_file_failures_do_not_skip_other_targets() {
-    static WARNINGS: Mutex<Vec<String>> = Mutex::new(Vec::new());
-    WARNINGS.lock().expect("lock warnings").clear();
-    struct RecordingReporter;
-    impl Reporter for RecordingReporter {
-        fn emit(event: &LogEvent) {
-            if let LogEvent::Pnpm(PnpmLog { level: LogLevel::Warn, message, .. }) = event {
-                WARNINGS.lock().expect("lock warnings").push(message.clone());
-            }
-        }
-    }
-
-    let dir = tempfile::tempdir().expect("create temp dir");
-    let pnpm_home_dir = dir.path().join("pnpm-home");
-    let bin_dir = pnpm_home_dir.join("bin");
-    let github_env = dir.path().join("a".repeat(300));
-    let github_path = dir.path().join("github-path");
-    std::fs::write(&github_path, "").expect("create github path");
-
-    persist_github_actions_environment_to_files::<RecordingReporter>(
-        dir.path(),
-        true,
-        &pnpm_home_dir,
-        &bin_dir,
-        Some(&github_env),
-        Some(&github_path),
-    );
-
-    assert_eq!(
-        std::fs::read_to_string(github_path).expect("read github path"),
-        format!("{}\n", bin_dir.display()),
-    );
-    let warnings = WARNINGS.lock().expect("lock warnings");
-    assert_eq!(warnings.len(), 1);
-    eprintln!("captured warning: {}", warnings[0]);
-    assert!(warnings[0].contains("GITHUB_ENV"));
-    assert!(warnings[0].contains(&github_env.display().to_string()));
-}
-
-#[test]
-fn github_actions_environment_files_are_not_created() {
-    let dir = tempfile::tempdir().expect("create temp dir");
-    let pnpm_home_dir = dir.path().join("pnpm-home");
-    let bin_dir = pnpm_home_dir.join("bin");
-    let github_env = dir.path().join("missing-github-env");
-
-    persist_github_actions_environment_to_files::<SilentReporter>(
-        dir.path(),
-        true,
-        &pnpm_home_dir,
-        &bin_dir,
-        Some(&github_env),
-        None,
-    );
-
-    assert!(!github_env.exists());
-}
-
-#[test]
-fn github_actions_environment_file_values_reject_line_breaking_characters() {
-    let err = validate_github_actions_environment_file_value(
-        "PNPM_HOME",
-        &PathBuf::from("/tmp/pnpm-home\nINJECTED=value"),
-    )
-    .expect_err("reject newline");
-
-    assert_eq!(err.to_string(), "PNPM_HOME cannot contain newline or NUL characters");
-    assert_eq!(
-        err.code().expect("diagnostic code").to_string(),
-        "ERR_PNPM_BAD_GITHUB_ACTIONS_ENVIRONMENT_VALUE",
-    );
 }
