+++ pnpm11/engine/pm/commands/src/setup/setup.ts
@@ -1,11 +1,9 @@
 import { spawnSync } from 'node:child_process'
 import fs from 'node:fs'
 import path from 'node:path'
-import util from 'node:util'
 
 import { detectIfCurrentPkgIsExecutable, packageManager } from '@pnpm/cli.meta'
 import { docsUrl } from '@pnpm/cli.utils'
-import { PnpmError } from '@pnpm/error'
 import { logger } from '@pnpm/logger'
 import {
   addDirToEnvPath,
@@ -14,6 +12,11 @@ import {
 } from '@pnpm/os.env.path-extender'
 import PATH from 'path-name'
 import { renderHelp } from 'render-help'
+
+import {
+  validateGitHubActionsEnvironmentFileValues,
+  writeGitHubActionsEnvironmentFiles,
+} from './githubActionsEnv.js'
 
 export const rcOptionsTypes = (): Record<string, unknown> => ({})
 
@@ -138,81 +141,6 @@ function createShellScript (targetDir: string, name: string, command: string): v
     fs.writeFileSync(path.join(targetDir, `${name}.cmd`), `@echo off\n${command} %*\n`)
     fs.writeFileSync(path.join(targetDir, `${name}.ps1`), `${command} @args\n`)
   }
-}
-
-function validateGitHubActionsEnvironmentFileValues (pnpmHomeDir: string, binDir: string): void {
-  if (!shouldPersistGitHubActionsEnvironmentFiles()) return
-  validateGitHubActionsEnvironmentFileValue('PNPM_HOME', pnpmHomeDir)
-  validateGitHubActionsEnvironmentFileValue('pnpm setup bin directory', binDir)
-}
-
-/**
- * `GITHUB_ENV` and `GITHUB_PATH` are line-oriented, so a line break in a
- * persisted value would append attacker-chosen records to the environment of
- * every later step in the workflow job.
- */
-function validateGitHubActionsEnvironmentFileValue (name: string, value: string): void {
-  if (value.includes('\n') || value.includes('\r') || value.includes('\0')) {
-    throw new PnpmError('BAD_GITHUB_ACTIONS_ENVIRONMENT_VALUE', `${name} cannot contain newline or NUL characters`)
-  }
-}
-
-function writeGitHubActionsEnvironmentFiles (pnpmHomeDir: string, binDir: string): void {
-  if (!shouldPersistGitHubActionsEnvironmentFiles()) return
-  const githubEnv = process.env.GITHUB_ENV
-  const githubPath = process.env.GITHUB_PATH
-  if (githubEnv != null) {
-    appendGitHubActionsEnvironmentFile('GITHUB_ENV', githubEnv, `PNPM_HOME=${pnpmHomeDir}`)
-  }
-  if (githubPath != null) {
-    appendGitHubActionsEnvironmentFile('GITHUB_PATH', githubPath, binDir)
-  }
-}
-
-function shouldPersistGitHubActionsEnvironmentFiles (): boolean {
-  return process.env.GITHUB_ACTIONS === 'true' && (process.env.GITHUB_ENV != null || process.env.GITHUB_PATH != null)
-}
-
-/**
- * The runner creates both files up front, so anything but an existing regular
- * file at `filePath` is not the runner's target: skip it instead of creating
- * it or following a symlink to it. A failure on one target must not stop the
- * other from being written.
- */
-function appendGitHubActionsEnvironmentFile (targetName: string, filePath: string, line: string): void {
-  try {
-    if (!fs.lstatSync(filePath).isFile()) return
-    appendLineToRegularFile(filePath, line)
-  } catch (err: unknown) {
-    if (util.types.isNativeError(err) && (err as NodeJS.ErrnoException).code === 'ENOENT') return
-    logger.warn({
-      message: `Failed to write GitHub Actions environment file ${targetName} (${filePath}): ${util.types.isNativeError(err) ? err.message : String(err)}`,
-      prefix: process.cwd(),
-    })
-  }
-}
-
-function appendLineToRegularFile (filePath: string, line: string): void {
-  const fd = fs.openSync(
-    filePath,
-    fs.constants.O_RDWR |
-      fs.constants.O_APPEND |
-      (process.platform === 'win32' ? 0 : fs.constants.O_NOFOLLOW)
-  )
-  try {
-    const stats = fs.fstatSync(fd)
-    if (!stats.isFile()) return
-    fs.writeSync(fd, `${missingRecordSeparator(fd, stats.size)}${line}\n`, null, 'utf8')
-  } finally {
-    fs.closeSync(fd)
-  }
-}
-
-function missingRecordSeparator (fd: number, size: number): string {
-  if (size === 0) return ''
-  const lastByte = Buffer.allocUnsafe(1)
-  const bytesRead = fs.readSync(fd, lastByte, 0, 1, size - 1)
-  return bytesRead === 1 && lastByte[0] !== 0x0A ? '\n' : ''
 }
 
 // v10-layout shim names that v11 writes under pnpmHomeDir/bin instead.
